Update realtime.py
implement median_of_three logistic measurement.
With this, N_BIN *should* be robust down to 9 bins- not that in principle anyone would ever use a voice channel that small, or should. In practice this will make ordinary speech channels more robust to noise, and lessen the impact of individual time segments with huge single outliers on the approach.